Consolidate *CompletionSources logic in System.IO.Pipes Common cancellation and completion logic moved into base class.
